Add the library generation as options in add-component command
