+++ dev/google-cloud
@@ -25,7 +25,6 @@ require __DIR__ . '/vendor/autoload.php';
 use Google\Cloud\Dev\Command\AddComponentCommand;
 use Google\Cloud\Dev\Command\ComponentInfoCommand;
 use Google\Cloud\Dev\Command\DocFxCommand;
-use Google\Cloud\Dev\Command\GenerateCloudLibrary;
 use Google\Cloud\Dev\Command\RepoInfoCommand;
 use Google\Cloud\Dev\Command\ReleaseInfoCommand;
 use Google\Cloud\Dev\Command\SplitCommand;
@@ -50,5 +49,4 @@ $app->add(new RepoInfoCommand());
 $app->add(new ReleaseInfoCommand());
 $app->add(new SplitCommand($rootDirectory));
 $app->add(new UpdateDepsCommand());
-$app->add(new GenerateCloudLibrary($rootDirectory));
 $app->run();
